Install Graphviz dependencies on Heroku-16
Some Graphviz-related dependencies are not present on the Heroku-16
stack (they are however present on the Cedar-14 stack). These
dependencies are now explicitly installed if the current stack is
Heroku-16.

Before this change, running a Graphviz binary (e.g. dot) on Heroku-16
resulted in a "shared library not found" error.