Add support for the Heroku-18 stack
On Ubuntu 18 we need to install `libgts` in addition to the shared
libraries required on Ubuntu 16. The library path changed so we add both
of them to `LD_LIBRARY_PATH`.

Added a small bash script to test the pack on Heroku-16 and Heroku-18
using their docker images. Sadly there is no image available for
Cedar-14.